+++ src/gridscorelb.jl
@@ -32,7 +32,7 @@ function gridscorelb(Xtrain, Ytrain, X, Y; score, fun, lb,
         ncomb = length(listpar)  # nb. combinations in pars
         res = list(ncomb)
         verbose ? println("-- Nb. combinations = ", ncomb) : nothing
-        for i = 1:ncomb
+        @inbounds for i = 1:ncomb
             verbose ? println(convert(NamedTuple, dat[i, :])) : nothing 
             listpar[i].lb = maximum(lb)
             fm = fun(Xtrain, Ytrain; par = listpar[i])